requirements.txt
absl-py==1.2.0
astunparse==1.6.3
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
confection==0.0.1
cymem==2.0.6
Flask==2.2.2
Flask-Markdown==0.3
Flask-WTF==1.0.1
flatbuffers==2.0.7
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.0
h5py==3.7.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
langcodes==3.3.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
murmurhash==1.0.8
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pathy==0.6.2
preshed==3.0.7
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
smart-open==5.2.1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.4
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-intel==2.10.0
tensorflow-cpu==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
thinc==8.0.17
tqdm==4.64.1
typer==0.4.2
typing-extensions==4.3.0
urllib3==1.26.12
wasabi==0.10.1
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.1
